+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Changelog
 New features:
 - Provide `do_fish_indent` TextCommand for indenting and prettifying the current file with `fish_indent` (#16)
   - Default shortcut to run the command is `ctrl+alt+i`
+- Provide a fish build system for running `do_fish_indent`
 
 2.2.6
 -----